requirements.txt
requests
py-cord
dadjokes
box
aiohttp
Pillow
pytz
yt-dlp
feedparser
python-dotenv
cloudscraper